moved away from android.app.Fragment in favor of support.v4.app.Fragment
