+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderPostListActivity.java
@@ -1,11 +1,11 @@
 package org.wordpress.android.ui.reader;
 
-import android.app.Fragment;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CoordinatorLayout;
+import android.support.v4.app.Fragment;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -159,7 +159,7 @@ public class ReaderPostListActivity extends AppCompatActivity {
             return;
         }
         Fragment fragment = ReaderPostListFragment.newInstanceForTag(tag, listType);
-        getFragmentManager()
+        getSupportFragmentManager()
                 .beginTransaction()
                 .replace(R.id.fragment_container, fragment, getString(R.string.fragment_tag_reader_post_list))
                 .commit();
@@ -175,7 +175,7 @@ public class ReaderPostListActivity extends AppCompatActivity {
             return;
         }
         Fragment fragment = ReaderPostListFragment.newInstanceForBlog(blogId);
-        getFragmentManager()
+        getSupportFragmentManager()
                 .beginTransaction()
                 .replace(R.id.fragment_container, fragment, getString(R.string.fragment_tag_reader_post_list))
                 .commit();
@@ -192,7 +192,7 @@ public class ReaderPostListActivity extends AppCompatActivity {
             return;
         }
         Fragment fragment = ReaderPostListFragment.newInstanceForFeed(feedId);
-        getFragmentManager()
+        getSupportFragmentManager()
                 .beginTransaction()
                 .replace(R.id.fragment_container, fragment, getString(R.string.fragment_tag_reader_post_list))
                 .commit();
@@ -205,7 +205,7 @@ public class ReaderPostListActivity extends AppCompatActivity {
     }
 
     private ReaderPostListFragment getListFragment() {
-        Fragment fragment = getFragmentManager().findFragmentByTag(getString(R.string.fragment_tag_reader_post_list));
+        Fragment fragment = getSupportFragmentManager().findFragmentByTag(getString(R.string.fragment_tag_reader_post_list));
         if (fragment == null) {
             return null;
         }
