+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderPostWebViewCachingFragment.java
@@ -1,9 +1,9 @@
 package org.wordpress.android.ui.reader;
 
 
-import android.app.Fragment;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -72,7 +72,7 @@ public class ReaderPostWebViewCachingFragment extends Fragment {
 
     private void selfRemoveFragment() {
         if (isAdded()) {
-            getActivity().getFragmentManager()
+            getActivity().getSupportFragmentManager()
                          .beginTransaction()
                          .remove(ReaderPostWebViewCachingFragment.this)
                          .commitAllowingStateLoss(); // we don't care about state here
